generate: Add support for multiple modules having same resource names
What we'll do is check if there are more than 1 module with resources, if so then we'll use full
resource name 'module.NAME.aws_instance.NAME2' as canonical. This will also be used on the config
output as key canonical for the resource.

If there is only 1 module they'll not be prefixed in favor of readability of the names.