cmd: Added new flag --description-file
Which will return a description file of the Nodes, which will contain
all the information from the read configuration.

We had to change the generate.FromHCL function to return now a new
parameter, so now it's the same as the generate.FromState.

Also the returned cofig from both of those methods has changed formats
as it's now just the resources on the first level without being them
nested inside a 'resources' and then grouped by resource type. It was
a newsense so we changed that